requirements.txt
asgiref==3.7.2
Django==4.2.5
sqlparse==0.4.4

djangorestframework~=3.14.0
Authlib~=1.2.1
django_redis~=5.3.0
whitenoise~=6.5.0